Some Sentimental Analysis creation is added in the system

+++ robot-model/src/main/java/com/share/investment/model/dao/SentimentAnalysis.java
@@ -33,7 +33,7 @@ public class SentimentAnalysis {
     @Column(
             nullable = false
     )
-    private byte score;
+    private Float score;
 
     @UpdateTimestamp
     @Temporal(TemporalType.TIMESTAMP)
